// src/app-public/modules/job-central/core/model/benchcandidate.ts
export class BenchCandidate {
    CandidateID: number;
    FirstName: string;
    MiddleName: string;
    LastName: string;
    Email:string;
    CandidateEmail: string;
    PrimaryPhoneNumber: string;
    Title: string;
    ResumePathBucket:string;
    ResumePathKey: string;
    Resume_Path_Type: string;
    ExperienceYears: string;
    City: string;
    State: string;
    WorkStatus: string;
    CreatedDate: Date;
    CreatedBy: string;
    BenchCreatedDate: Date;
    LinkedIn: string;
    BenchCategory: string;
    EmployeeID: number;
    Assignments: number;
    ViewResumeInnerPath: string;
    Certifications: string;
    AvailabilityToStart: string;
    Skillset: string;
    HashTags: string;
    Rate: string;
    BillingType: number;
    CreatedEmployee: string;
    ResumeSource: string;
    BenchCategoryName: string;
    Coordinator: string;
    BenchCandidate: boolean;
    SubmissionsCount: number;
    InterviewsCount: number;
    Location: string;
    labels: string[];
    SellRate: string;
    SellRateType: number;
    WorkStatusExpiry: Date;
    Source: string;
    UpdatedBy: string;
    UpdatedDate: Date;
    StatusName: string;
    WillingtoWork: string;
    Assigned_To: string[];
    CompanyId: number;
    Status: number;
    BenchPriority: string;
    UploadedFileName: string;
    TempKey: string;
    Qualification: string;
    AssignedList: any[];
    DisplayFileName: string;
    ResumePathType: string;
    candoptions:string;
    NotesCount:number;
    LastUpdated:Date;
    constructor() {
        this.CandidateID = 0;
        this.FirstName = null;
        this.LastName = null;
        this.Email=null;
        this.CandidateEmail = null;
        this.PrimaryPhoneNumber = null;
        this.Title = null;
        this.ResumePathBucket = null;
        this.ResumePathKey = null;
        this.Resume_Path_Type = null;
        this.ExperienceYears = null;
        this.City = null;
        this.State = null;
        this.WorkStatus = null;
        this.CreatedDate = null;
        this.CreatedBy = null;
        this.BenchCreatedDate = null;
        this.LinkedIn = null;
        this.BenchCategory = null;
        this.EmployeeID = null;
        this.Assignments = null;
        this.ViewResumeInnerPath = null;
        this.Certifications = null;
        this.AvailabilityToStart = null;
        this.Skillset = null;
        this.HashTags = null;
        this.Rate = null;
        this.BillingType = null;
        this.CreatedEmployee = null;
        this.ResumeSource = null;
        this.BenchCategoryName = null;
        this.Coordinator = null;
        this.BenchCandidate = false;
        this.SubmissionsCount = 0;
        this.InterviewsCount = 0;
        this.Location = null;
        this.labels = [];
        this.WorkStatusExpiry = new Date();
        this.Source = null;
        this.UpdatedBy = null;
        this.UpdatedDate = null;
        this.StatusName = null;
        this.WillingtoWork = null;
        this.Assigned_To = [];
        this.CompanyId = 0;
        this.Status = 0;
        this.BenchPriority = null;
        this.UploadedFileName = null;
        this.TempKey = null;
        this.Qualification = null;
        this.AssignedList = [];
        this.DisplayFileName = null;
        this.ResumePathType = null;
        this.candoptions=null;
        this.NotesCount=0;
        this.LastUpdated = null;
    }
}